+++ apps/teams-test-app/src/components/AuthenticationAPIs.tsx
@@ -51,6 +51,9 @@ const GetAuthToken = (): React.ReactElement =>
         },
       },
     },
+    defaultInput: JSON.stringify({
+      resources: ['https://teams.microsoft.com'],
+    }),
   });
 
 const GetUser = (): React.ReactElement =>
@@ -131,7 +134,7 @@ const Authenticate = (): React.ReactElement =>
         },
       },
     },
-    defaultInput: '{"url":"https://www.microsoft.com"}',
+    defaultInput: '{"url":"https://localhost:4000"}',
   });
 
 const getAuthParams = (authParam: authAuthenticateParams): authentication.AuthenticatePopUpParameters => {