+++ apps/teams-test-app/src/pages/TestApp.tsx
@@ -5,6 +5,7 @@ import AppAPIs from '../components/AppAPIs';
 import AppEntityAPIs from '../components/AppEntityAPIs';
 import AppInitializationAPIs from '../components/AppInitialization';
 import AppInstallDialogAPIs from '../components/AppInstallDialog';
+import AppPerformanceMetricsAPIs from '../components/AppPerformanceMetrics';
 import AuthenticationAPIs from '../components/AuthenticationAPIs';
 import BarCodeAPIs from '../components/BarCodeAPIs';
 import CalendarAPIs from '../components/CalendarAPIs';
@@ -54,11 +55,13 @@ import NotificationAPIs from '../components/privateApis/NotificationAPIs';
 import PrivateAPIs from '../components/privateApis/PrivateAPIs';
 import TeamsAPIs from '../components/privateApis/TeamsAPIs';
 import VideoExAPIs from '../components/privateApis/VideoEffectsExAPIs';
+import WidgetHostingAPIs from '../components/privateApis/WidgetHostingAPIs';
 import ProfileAPIs from '../components/ProfileAPIs';
 import RemoteCameraAPIs from '../components/RemoteCameraAPIs';
 import SearchAPIs from '../components/SearchAPIs';
 import SecondaryBrowserAPIs from '../components/SecondaryBrowserAPIs';
 import SharingAPIs from '../components/SharingAPIs';
+import ShortcutRelayAPIs from '../components/ShortcutRelayAPIs';
 import StageViewAPIs from '../components/StageViewAPIs';
 import StageViewSelfAPIs from '../components/StageViewSelfAPIs';
 import StoreAPIs from '../components/StoreApis';
@@ -100,6 +103,7 @@ export const TestApp: React.FC = () => {
       { name: 'AppAPIs', component: <AppAPIs /> },
       { name: 'AppInitializationAPIs', component: <AppInitializationAPIs /> },
       { name: 'AppInstallDialogAPIs', component: <AppInstallDialogAPIs /> },
+      { name: 'AppPerformanceMetricsAPIs', component: <AppPerformanceMetricsAPIs /> },
       { name: 'AuthenticationAPIs', component: <AuthenticationAPIs /> },
       { name: 'AppEntityAPIs', component: <AppEntityAPIs /> },
       { name: 'BarCodeAPIs', component: <BarCodeAPIs /> },
@@ -157,6 +161,7 @@ export const TestApp: React.FC = () => {
       { name: 'SearchAPIs', component: <SearchAPIs /> },
       { name: 'SecondaryBrowserAPIs', component: <SecondaryBrowserAPIs /> },
       { name: 'SharingAPIs', component: <SharingAPIs /> },
+      { name: 'ShortcutRelayAPIs', component: <ShortcutRelayAPIs /> },
       { name: 'WebStorageAPIs', component: <WebStorageAPIs /> },
       { name: 'StageViewAPIs', component: <StageViewAPIs /> },
       { name: 'StageViewSelfAPIs', component: <StageViewSelfAPIs /> },
@@ -167,6 +172,7 @@ export const TestApp: React.FC = () => {
       { name: 'VideoExAPIs', component: <VideoExAPIs /> },
       { name: 'VisualMediaAPIs', component: <VisualMediaAPIs /> },
       { name: 'StoreAPIs', component: <StoreAPIs /> },
+      { name: 'WidgetHostingAPIs', component: <WidgetHostingAPIs /> },
     ],
     [],
   );