+++ packages/teams-js/src/public/runtime.ts
@@ -230,6 +230,7 @@ interface IRuntimeV4 extends IBaseRuntime {
     };
     readonly appEntity?: {};
     readonly appInstallDialog?: {};
+    readonly appPerformanceMetrics?: {};
     readonly barCode?: {};
     readonly calendar?: {};
     readonly call?: {};
@@ -240,6 +241,7 @@ interface IRuntimeV4 extends IBaseRuntime {
       readonly customTelemetry?: {};
       readonly eligibility?: {};
       readonly sidePanel?: {};
+      readonly view?: {};
     };
     readonly dialog?: {
       readonly card?: {
@@ -297,6 +299,7 @@ interface IRuntimeV4 extends IBaseRuntime {
     readonly sharing?: {
       readonly history?: {};
     };
+    readonly shortcutRelay?: {};
     readonly stageView?: {
       readonly self?: {};
     };
@@ -316,6 +319,7 @@ interface IRuntimeV4 extends IBaseRuntime {
       readonly image?: {};
     };
     readonly webStorage?: {};
+    readonly widgetHosting?: {};
   };
 }
 // Constant used to set the runtime configuration